Moves GPSButton into AssetSelect components folder
